pkg/util: move CompileRegex helper into apiserver
This is a little helper, not a generic utility of general value.
